+++ src/pages/PostJobs/FormSections/index.ts
@@ -1,0 +1,8 @@
+export { default as JobTitleSection } from './JobTitleSection';
+export { default as CompanyDetailsSection } from './CompanyDetailsSection';
+export { default as CompanyLocationSection } from './CompanyLocationSection';
+export { default as JobDates } from './JobDates';
+export { default as QualificationsSection } from './QualificationsSection';
+export { default as DutiesSection } from './DutiesSection';
+export { default as SalarySection } from './SalarySection';
+export { default as SubmitSection } from './SubmitSection';
