+++ src/pages/PostJobs/validate.ts
@@ -1,0 +1,14 @@
+import  FormData  from '../../components/DataModels/FormData';
+
+const validate=(form:FormData)=>{
+  const mandatoryFields = [
+    'title', 'qualification', 'experience', 'companyName', 'companyType', 'companyLogo', 'city', 'state', 'country',
+    'region', 'postingDate', 'expiryDate', 'appClosingDate', 'removingJobDate', 'salary', 'hours', 'jobType'
+  ];
+  const isValid = mandatoryFields.reduce((acc,field) => { return  acc && !form[field]; },true);
+
+  return isValid ? null: form;
+
+};
+
+export default validate;
